Materialis

Documentation for Materialis.

Materialis.AbstractMeasuredData

Materialis.compute_disps

Materialis.image_based_identification

Materialis.nodes2dofs

Materialis.ranges

Materialis.vtkStrGridPlot

Materialis.vtkUnsGridPlot

Materialis.AbstractMeasuredData — Type

AbstractMeasuredData

Abstract supertype that holds the measured data information.

Materialis.compute_disps — Method

Function that computes the displacement using external user-provided tools at the points of the image grid.

Materialis.image_based_identification — Method

matParams: vector with material parameters

Materialis.nodes2dofs — Method

copied from https://github.com/jorgepz/FEMAssembler.jl/blob/main/src/mesh/dofsComputations.jl

Materialis.ranges — Method

funcion para calculo rangos de indices xyz de grid que efecitivamente caen dentro del box del elemento

Materialis.vtkStrGridPlot — Method

vtkStrGridPlot: function for writing vtk structured grid files and magnitudes

nodalMagnitudes: vector of Any with [ magnitudes1, magnitudes2, ...]

Materialis.vtkUnsGridPlot — Method

vtkUnsGridPlot: function for writing vtk unstructured grid files and magnitudes